Photography-Website
The Malcolm-Lismore-Photography-Website project is a comprehensive photography website designed for Malcolm Lismore. It features several key pages and functionalities, including a home page showcasing the best work and an introduction to Malcolm Lismore's photography, a user registration system for new users to sign up and existing users to log in, a works gallery displaying the portfolio of photography categorized for easy browsing, an integrated booking system enabling clients to schedule photography sessions, detailed information about various photography packages offered, and a contact page for users to reach out with inquiries or to get in touch directly. This project aims to provide a seamless and professional experience for users, highlighting Malcolm's photography services and facilitating easy interaction and bookings. Please review the code and feel free to suggest any improvements or report issues.